requirements.txt
requests==2.25.1
Flask~=1.1.2
PyExecJS==1.5.1